Computer control with Open-Cv
The thumb, index finger, middle finger, and pinky of your hand are used to increase or decrease the computer's volume and pause the video. If you touch your thumb and index finger together, the volume increases. If you touch your thumb and middle finger together, the volume decreases. If you touch your thumb and pinky together, the video pauses.